week4: (task 2) create Manager class overriding Employee

+++ src/oop_00000113459_MuhammadDanielAzmi/week4/Employee.kt
@@ -1,6 +1,6 @@
 package oop_00000113459_MuhammadDanielAzmi.week4
 
-open class Employee(val name: String, val baseSalary: Int) {
+open class Employee(val name: String, val baseSalary: Double) {
 
     open fun work() {
         println("Employee $name sedang bekerja.")
